refactor(adapter-openclaw): extract agent-turn I/O into agent-turn.ts
Tenth cut of the runtime.ts god-file split. Moves the 5 pure agent-turn I/O
helpers into a new agent-turn.ts: messagesToOpenClawPrompt (message-list →
CLI prompt), normalizeToolList (tool allow-list dedup), oversizedOutputBytesFrom,
and extractOpenClawAgentText / extractOpenClawAgentUsage (assistant text +
token usage out of the gateway/CLI turn response). Pure parsing/formatting
over the turn payload.

applyRuntimeMessageToolPolicy stays on the adapter — it is typed against
OpenClawAgentTurnOptions (a class-central type, 8 uses, depends on MessageArgs);
it imports normalizeToolList back. The class's turn methods import back all 5.

Pure relocation — bodies byte-for-byte unchanged. runtime.ts 1,618 -> 1,560,
shed its now-unused getJsonPath/isPlainObject imports.

This completes the runtime.ts free-function extraction: 3,188 -> 1,560 (51%)
across 10 stacked PRs into 10 cohesive sibling modules (runtime-utils, errors,
cron-store, activity-summary, image-inference, approval-helpers, session-activity,
agent-config, channel-helpers, agent-turn). What remains is the
OpenClawRuntimeAdapter capability facade + its class-central types
(OpenClawSettings, OpenClawAgentTurnOptions) + the two type-coupled glue fns
(gatewayWebSocketUrl, applyRuntimeMessageToolPolicy) + the settings/binary
resolver. The capability-subclass refactor of the facade itself is
behavior-touching (Tier 3) and deliberately not taken.

Verified: typecheck, eslint (changed files), adapter suite 115-0 (--isolate),
full suite 5124-0, binary build (stamps reverted), boot smoke (adapter loads,
schedule activates, list route 200).
